Use integer types in reinterpret-of-view test for portability
Float16/Float32/Float64 are not supported on all GPU backends, so use
Int32/Int64 instead so the test runs everywhere.

+++ test/testsuite/base.jl
@@ -443,10 +443,10 @@ end
       @test collect(reinterpret(Int32, AT(fill(1f0))))[] == reinterpret(Int32, 1f0)
 
       @testset "reinterpret of view with non-aligned offset" begin
-        a = AT(Float16[1,2,3,4,5,6,7,8,9])
-        v = view(a, 2:7)  # offset of 1 Float16 = 2 bytes
-        r = reinterpret(Float32, v)  # Float32 = 4 bytes; 2 is not a multiple of 4
-        @test Array(r) == reinterpret(Float32, @view Array(a)[2:7])
+        a = AT(Int32[1,2,3,4,5,6,7,8,9])
+        v = view(a, 2:7)  # offset of 1 Int32 = 4 bytes
+        r = reinterpret(Int64, v)  # Int64 = 8 bytes; 4 is not a multiple of 8
+        @test Array(r) == reinterpret(Int64, @view Array(a)[2:7])
       end
 
       @testset "reinterpret(reshape)" begin
